# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/hotel/store": {
            "post": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Create of hotel",
                "description": "Create of hotels",
                "operationId": "hotelStore",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number_rooms",
                        "in": "query",
                        "description": "Number Rooms of the hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nit",
                        "in": "query",
                        "description": "Nit Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotel/update/{id}": {
            "put": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Update hotel",
                "description": "Update hotel",
                "operationId": "hotelUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "City Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number_rooms",
                        "in": "query",
                        "description": "Number Rooms of the hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Address Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nit",
                        "in": "query",
                        "description": "Nit Hotel",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotel/show/{id}": {
            "get": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Read of hotel",
                "description": "Read of hotels",
                "operationId": "hotelGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotel/index": {
            "get": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Read all the hotels",
                "description": "Read all the hotels",
                "operationId": "hotelAll",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotel/delete/{id}": {
            "delete": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Delete a hotel",
                "description": "Delete a hotel",
                "operationId": "hoteldelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hotel/room/store": {
            "post": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Create type room hotel",
                "description": "Create type room hotel",
                "operationId": "storeRoomHotel",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id Hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type_room_id",
                        "in": "query",
                        "description": "Id Type Room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "accommodation_id",
                        "in": "query",
                        "description": "Id Type Accommodation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Number Rooms of the type room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}